Added control word on content module deletes

+++ modules/content/actions/delete-file.php
@@ -9,21 +9,26 @@
 
 defined('ANS') or die();
 
+if ($Vars->var['confirm'] !== __('DELETE')) {
+    $Vars->message(__('Please write exact "DELETE" if you want to delete this content'), 'error');
+    return false;
+}
+
+$Vars->delete('confirm');
+
 $Files = new \ANS\PHPCan\Files\File;
 
 $file = filePath('scene/uploads|'.$Vars->get('file'));
 
 if (!is_file($file)) {
     $Vars->message(__('This file doesn\'t exits'), 'error');
-
     return false;
 }
 
 $ok = $Files->delete($file);
 
-if (!$ok) {
+if (empty($ok)) {
     $Vars->message(__('There was an error deleting the file'), 'error');
-
     return false;
 }
